Before Removal of .*SHW
>> taken from the discord server

hi all! i want to share an important update, not just for future users of duofur, but also for those who develop and play projects based on the rr engine.

over time, it has become clear that developing on the .shw format (rshw, cshw, etc.) is not sustainable. there are numerous issues with the format, but more importantly, it poses a serious security risk. the format has long been known - even by 64th - to rely on the dangerous BinaryFormatter class. while BinaryFormatter makes it easy to serialize and deserialize objects, that flexibility is also its greatest weakness.

the core vulnerability lies in deserialization. a malicious actor could craft a file containing harmful code, which the engine would treat like any other showtape. if executed successfully, this could allow an attacker to run arbitrary code through duofur’s thread, posing a severe security risk. given how dangerous this is, i’ve decided to deprecate the format and replace it with a new, open-source format that anyone can modify and implement in their own projects.  while playback support for .shw files will remain (along with a converter), there will be no more recording functionality for the format. i understand this change may be disappointing for some, but keeping people's data safe is important

this new format will likely use JSON instead of BinaryFormatter, this has the possibility of making it human-readable, which means it's easier to debug, and more open! i encourage all mod developers to consider adding this new format as an option - it'll make life easier for everyone!

thank you for understanding!